requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
uuid==1.30
